pytest: Migrate payment related tests to use modern fixtures
